Switched co-ordinate type, to line up with new absolute func

+++ 2017/03/02/day3_part2_test.go
@@ -61,7 +61,7 @@ func TestLastValue(t *testing.T) {
 func TestManhattan(t *testing.T) {
 	cases := []struct {
 		in   int
-		want uint
+		want uint64
 	}{
 		{1, 0},
 		{12, 3},
